feat: split aios product surface from runtime admin
