+++ docs/backlog.md
@@ -155,7 +155,9 @@ Status Operating Cycle deploy real em 2026-05-07: CT165 atualizado ate `e518f3d`
 
 Status Production Operating Loop v0 em 2026-05-07: implementado e deployado runner recorrente interno no daemon, configuravel por `AIOS_COMPANY_BRAIN_OPERATING_LOOP_*`, sem cron/systemd timer novo e sem writeback externo. O loop checa watchers due/stale, roda somente `watcher-aios-briefing-v0` e `watcher-github-pr-ci-v0` com `actionPolicy=observe_only`, usa lock em memoria e grava provenance `schedule://...`. Exposto em `/api/company-brain/operating-loop`, `operatingCadence.operatingLoop`, `coreReadiness.operatingLoop`, Operating UI e MCP `get_company_brain_operating_loop`. Dogfood local validou run automatico com `watcherRunsCreated=2`, `artifactsCreated=2`, `activeScheduledWatcherCount=2`, `staleCadenceCount=0`, `dueCadenceCount=0` e `dailyUseBlockingGapCount=0`. Em producao CT165, primeiro tick automatico rodou `watcher-github-pr-ci-v0` com `triggerRef=schedule://production%3Acompany-brain-operating-loop/DB8t6HVmDsX-`, `artifactId=KX7Ujpfv80MO`, `activeScheduledWatcherCount=2`, `staleCadenceCount=0`, `dueCadenceCount=0`, `dailyUseBlockingGapCount=0` e Operating Snapshot `5/5 operating cards ready`. Segundo tick confirmou recorrencia: `tickCount=2`, `runCount=1`, `skippedTickCount=1`, `lastDueWatcherIds=[]`, sem erro.
 
-Proximo corte planejado: usar `https://ai.felhen.ai/company-brain/operating` como ciclo diario persistente e coletar friccoes reais antes de criar features novas. Continuar automaticamente apenas para read-only/audit/docs/UI/MCP/dogfood; parar antes de novo executor real, novo alvo externo, secrets/permissoes ou qualquer mutacao fora da policy aprovada.
+Status AIOS GitHub Roadmap Reset em 2026-05-07: issues antigas `#1` a `#6` foram triadas como legado/icebox/superseded e fechadas como `not planned`, sem apagar contexto. Criada milestone `AIOS Execution Loop v0`, labels de produto/runtime e fila ativa `#25` a `#31`: `AIOS-CLEAN-00`, `AIOS-CLEAN-01`, `AIOS-EXEC-01` a `AIOS-EXEC-05`. Sync read-only de GitHub Issues em producao criou Source `AIOS GitHub Issues active roadmap` e WorkItems canonicos para as 7 issues. Registro em `docs/action/aios-github-roadmap-reset-2026-05-07.md`.
+
+Proximo corte planejado: consumir `#25 AIOS-CLEAN-00: Product Surface Split` via branch/PR, separando AIOS produto de Runtime Admin legado antes de iniciar features do Execution Command Center. Continuar automaticamente apenas para read-only/audit/docs/UI/MCP/dogfood; parar antes de novo executor real, novo alvo externo, secrets/permissoes ou qualquer mutacao fora da policy aprovada.
 
 - [x] **Company Brain schema v0** — adicionar objetos horizontais no daemon: `Source`, `Artifact`, `StrategicPriority`, `Decision`, `Signal`, `WorkItem`, `WorkflowBlueprint`, `WorkflowRun`, `AlignmentFinding`, `GuidanceItem`, `AgentContext` e `ImprovementProposal`
 - [x] **Source registry + raw artifact store** — guardar artifacts com `source`, `raw_ref`, author, timestamp, hash, visibility e provenance
@@ -168,6 +170,13 @@ Proximo corte planejado: usar `https://ai.felhen.ai/company-brain/operating` com
 - [x] **Demo/Readiness Cleanup v0** — reduzir friccoes do pack com proposals internas v0.2, Source Health menos ruidoso, alias top-level de replay, reconhecimento do design-partner pack e campo canonico `artifactType`
 - [x] **Operating Cycle Friction Closure v0** — corrigir briefing pos-cadence, `runs[].artifactId`, status agregado de Gate Closure/Operating Snapshot e semantica de Core Readiness antes de deployar o ciclo diario real
 - [x] **Production Operating Loop v0** — runner recorrente interno no daemon para executar watchers read-only due/stale sem sessao interativa aberta, com env config, lock, provenance, API/UI/MCP e dogfood local
+- [ ] **AIOS-CLEAN-00 Product Surface Split** — fazer AIOS Operating ser a entrada principal, preservar runtime antigo em `/runtime` e agrupar Console/Terminal/Jobs/Schedules/Costs/Inbox/Settings como Runtime Admin
+- [ ] **AIOS-CLEAN-01 GitHub Pipeline Hygiene** — manter issues GitHub como fila limpa e sincronizada ao Company Brain
+- [ ] **AIOS-EXEC-01 Execution Command Center v0** — recomendar proximo WorkItem e gerar prompt acionavel
+- [ ] **AIOS-EXEC-02 WorkItem to GitHub Issue Flow** — criar issue GitHub governada a partir de WorkItem/Guidance
+- [ ] **AIOS-EXEC-03 Agent Session Launcher** — gerar handoff/prompt por WorkItem e agente alvo
+- [ ] **AIOS-EXEC-04 Session Result Intake** — transformar resultado de sessao em Artifact/Signal/Finding/Guidance
+- [ ] **AIOS-EXEC-05 Project Pipeline View** — mostrar pipeline por projeto/repo/gate/PR
 - [x] **Watcher status em Source Health / Summary** — expor status basico de watchers, ultimas execucoes, erro, artifacts/signals/work items gerados e freshness em `/api/company-brain/summary` ou tela de Source Health
 - [x] **Watcher manual/simulado v0** — criar watcher manual para PR/CI ou GitHub Issues que consiga registrar uma execucao sem webhook real
 - [x] **Watcher output com provenance e policy** — garantir que uma execucao de watcher consiga gerar `Artifact`, `Signal` ou `WorkItem` com provenance, `action_policy`, `risk_class` e trilha auditavel antes de qualquer writeback. Implementado para `Artifact`, `Signal`, `WorkItem`, `AlignmentFinding` e `GuidanceItem`.
